+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/supplies/AcquisitionDto.kt
@@ -1,0 +1,8 @@
+package com.app.arcabyolimpo.data.remote.dto.supplies
+
+import com.google.gson.annotations.SerializedName
+
+data class AcquisitionDto(
+    @SerializedName("idTipoAdquisicion") val id: String,
+    @SerializedName("Descripcion") val description: String,
+)