+++ app/src/main/java/com/app/arcabyolimpo/data/repository/productbatches/ProductBatchRepositoryImpl.kt
@@ -2,6 +2,8 @@ package com.app.arcabyolimpo.data.repository.productbatches
 
 import com.app.arcabyolimpo.data.mapper.productbatches.toDomain
 import com.app.arcabyolimpo.data.mapper.productbatches.toDto
+import com.app.arcabyolimpo.data.mapper.productbatches.toModifyDto
+import com.app.arcabyolimpo.data.mapper.productbatches.toRegisterDto
 import com.app.arcabyolimpo.data.remote.api.ArcaApi
 import com.app.arcabyolimpo.domain.model.productbatches.ProductBatch
 import com.app.arcabyolimpo.domain.repository.productbatches.ProductBatchRepository
@@ -16,19 +18,21 @@ class ProductBatchRepositoryImpl(
 ) : ProductBatchRepository {
     override suspend fun getProductBatches(): List<ProductBatch> = api.getProductBatches().map { it.toDomain() }
 
-    override suspend fun getProductBatch(id: String): ProductBatch {
-        val dto = api.getProductBatch(id)
-        return dto.toDomain()
+    override suspend fun getProductBatch(id: String): ProductBatch = api.getProductBatch(id).toDomain()
+
+    override suspend fun registerProductBatch(batch: ProductBatch) {
+        val dto = batch.toRegisterDto()
+        api.addProductBatch(dto)
     }
 
-    //  override suspend fun getProductBatch(id: String): ProductBatch = api.getProductBatch(id).toDomain()
-
-    override suspend fun registerProductBatch(batch: ProductBatch): Result<Unit> =
-        try {
-            val dto = batch.toDto()
-            api.addProductBatch(dto)
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
+    override suspend fun modifyProductBatch(
+        batch: ProductBatch,
+        id: String,
+    ) {
+        val dto = batch.toModifyDto()
+        api.modifyProductBatch(
+            id = id,
+            batch = dto,
+        )
+    }
 }