+++ app/src/main/java/com/app/arcabyolimpo/domain/repository/supplies/SupplyRepository.kt
@@ -3,6 +3,10 @@ package com.app.arcabyolimpo.domain.repository.supplies
 import android.net.Uri
 import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
 import com.app.arcabyolimpo.domain.model.filter.FilterData
+import com.app.arcabyolimpo.domain.model.supplies.Acquisition
+import com.app.arcabyolimpo.domain.model.supplies.RegisterSupplyBatch
+import com.app.arcabyolimpo.domain.model.supplies.SuccessMessage
+import com.app.arcabyolimpo.domain.model.supplies.SupplyBatch
 import com.app.arcabyolimpo.domain.model.supplies.Batch
 import com.app.arcabyolimpo.domain.model.supplies.Supply
 import com.app.arcabyolimpo.domain.model.supplies.SupplyAdd
@@ -21,12 +25,16 @@ interface SupplyRepository {
 
     suspend fun getSupplyById(id: String): Supply
 
+    suspend fun registerSupplyBatch(batch: RegisterSupplyBatch): SuccessMessage
+
     suspend fun getSupplyBatchById(id: String): SupplyBatchExt
 
     suspend fun filterSupply(params: FilterDto): List<Supply>
 
     suspend fun getFilterData(): FilterData
 
+    suspend fun getAcquisitionTypes(): List<Acquisition>
+    
     suspend fun deleteSupplyBatch(id: String)
 
     // Yo uso Update