+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/productbatches/mapper/ProductBatchUiMapper.kt
@@ -2,6 +2,7 @@ package com.app.arcabyolimpo.presentation.screens.productbatches.mapper
 
 import com.app.arcabyolimpo.domain.model.productbatches.ProductBatch
 import com.app.arcabyolimpo.presentation.screens.productbatches.model.ProductBatchUiModel
+import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchModify.ProductBatchModifyUiState
 import com.app.arcabyolimpo.presentation.screens.productbatches.productBatchRegister.ProductBatchRegisterUiState
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
@@ -45,7 +46,25 @@ fun ProductBatchRegisterUiState.toDomain() =
         precioVenta = precioVenta,
         descripcion = "",
         disponible = 0,
-        cantidadProducida = 0,
+        cantidadProducida = cantidadProducida.toIntOrNull() ?: 0,
+        fechaRealizacion = fechaRealizacion,
+        fechaCaducidad = fechaCaducidad,
+        idInventario = "",
+        precioUnitario = "0.0",
+        imagen = "",
+    )
+
+/** Maps a [ProductBatchModifyUiState] UI state to a [ProductBatch] domain model for business logic.
+ *  @return ProductBatch
+ */
+fun ProductBatchModifyUiState.toDomain() =
+    ProductBatch(
+        idProducto = "",
+        nombre = "",
+        precioVenta = precioVenta,
+        descripcion = "",
+        disponible = 0,
+        cantidadProducida = cantidadProducida.toIntOrNull() ?: 0,
         fechaRealizacion = fechaRealizacion,
         fechaCaducidad = fechaCaducidad,
         idInventario = "",